+++ src/ddzy/utility/array/index.ts
@@ -11,6 +11,7 @@ export interface IUtilityArrayProps {
   isStrictArray(origin: any): boolean;
   toFlatArrayOutPlace(origin: any[]): any[];
   toStrictArray<T>(origin: ArrayLike<T>): T[];
+  trunk<I>(origin: I[], size?: number): I[][];
 };
 
 
@@ -73,6 +74,26 @@ const utilityArray: IUtilityArrayProps = {
         result.push(every);
       }
     }
+
+    return result;
+  },
+
+  /**
+   * 将源数组根据指定大小分片
+   * @param origin 源数组
+   * @param size 分片大小
+   */
+  trunk<I>(origin: I[], size = 1) {
+    const result: I[][] = [];
+    let count = 0;
+
+    if (!origin.length) {
+      return result;
+    }
+
+    do {
+      result.push(origin.slice(count, count + size));
+    } while ((count += size) < origin.length);
 
     return result;
   },